+++ src/Voting.Lib.VotingExports/Repository/Ausmittlung/AusmittlungPdfVoteTemplates.cs
@@ -120,6 +120,7 @@ public static class AusmittlungPdfVoteTemplates
         EntityType = EntityType.Vote,
         ResultType = ResultType.PoliticalBusinessResultBundleReview,
         GeneratedBy = VotingApp.VotingAusmittlung,
+        AsyncJobPriority = 10,
     };
 
     internal static readonly IReadOnlyCollection<TemplateModel> All = new[]